+++ qt-locate/src/tests/res/query_qt4.8.7.in
