+++ qt-locate/src/tests/res/query_qt5.11.1.in
