+++ qt-locate/src/tests/res/query_qt5_test_win.in
