+++ qt-locate/src/tests/res/query_qt_no_bin.in
