+++ qt-locate/src/tests/res/query_qt_no_include.in
