+++ tests/test-build/build.rs
@@ -1,17 +1,9 @@
-use qt_binding_build::{build::Builder, locate::locate, Version};
+use qt_binding_build::Builder;
 
 fn main() {
-    let qt_install = locate().unwrap();
-
-    match qt_install.major_version() {
-        Version::Qt5 => {
-            println!("cargo:rustc-cfg=qt5");
-
-            Builder::from_install(qt_install)
-                .res_file("src/res.qrc")
-                .moc_file("src/object.h")
-                .files(&["src/bindings.cpp", "src/object.cpp"])
-                .build("bindings");
-        }
-    }
+    Builder::new()
+        .res_file("src/res.qrc")
+        .moc_file("src/object.h")
+        .files(&["src/bindings.cpp", "src/object.cpp"])
+        .build("bindings");
 }
