+++ tests/test-locate/src/lib.rs
@@ -1,4 +1,4 @@
-//! `qt-binding-core` locate test
+//! `qt-sys` locate test
 //!
-//! This crate contains one test, that is used to test `qt-binding-core`'s capability to locate
+//! This crate contains one test, that is used to test `qt-locate`'s capability to locate
 //! Qt. This crate's test should run on a CI containing a Qt install.
